// src/shared/models/application.model.ts
export interface Application {
  id: string;
  createdAt: Date;
  updatedAt: Date;
  gender: string;
  name: string;
  fathersName: string;
  phone: string;
  permanentAddress: string;
  academyName: string;
  academyAddress: string;
  budget: string;
  expectedEntryDate: string;
  condition: boolean;
}